refactor(qemu): Migrate init method to its own file
Since the init method contains largely boilerplate for future gob
registrations, the method is quite large.  Migrate to its own file to
keep the driver file smaller and more focused.
