+++ machine/qemu/init.go
@@ -1,0 +1,451 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2022, Unikraft GmbH and The KraftKit Authors.
+// Licensed under the BSD-3-Clause License (the "License").
+// You may not use this file except in compliance with the License.
+package qemu
+
+import "encoding/gob"
+
+func init() {
+	// Register only used supported interfaces later used for serialization.  To
+	// include all will roughly increase the final binary size by +20MB.
+
+	// Character devices
+	// gob.Register(QemuCharDevNull{})
+	// gob.Register(QemuCharDevSocketTCP{})
+	// gob.Register(QemuCharDevSocketUnix{})
+	// gob.Register(QemuCharDevUdp{})
+	// gob.Register(QemuCharDevVirtualConsole{})
+	// gob.Register(QemuCharDevRingBuf{})
+	// gob.Register(QemuCharDevFile{})
+	// gob.Register(QemuCharDevPipe{})
+	// gob.Register(QemuCharDevPty{})
+	// gob.Register(QemuCharDevStdio{})
+	// gob.Register(QemuCharDevSerial{})
+	// gob.Register(QemuCharDevTty{})
+	// gob.Register(QemuCharDevParallel{})
+	// gob.Register(QemuCharDevParport{})
+	// gob.Register(QemuCharDevSpiceVMC{})
+	// gob.Register(QemuCharDevSpicePort{})
+
+	// Host character devices
+	// gob.Register(QemuHostCharDevVirtualConsole{})
+	// gob.Register(QemuHostCharDevPty{})
+	gob.Register(QemuHostCharDevNone{})
+	// gob.Register(QemuHostCharDevNull{})
+	// gob.Register(QemuHostCharDevNamed{})
+	// gob.Register(QemuHostCharDevTty{})
+	gob.Register(QemuHostCharDevFile{})
+	// gob.Register(QemuHostCharDevStdio{})
+	// gob.Register(QemuHostCharDevPipe{})
+	// gob.Register(QemuHostCharDevUDP{})
+	// gob.Register(QemuHostCharDevTCP{})
+	// gob.Register(QemuHostCharDevTelnet{})
+	// gob.Register(QemuHostCharDevWebsocket{})
+	gob.Register(QemuHostCharDevUnix{})
+
+	// CPU devices
+	// gob.Register(QemuDevice486V1X8664Cpu{})
+	// gob.Register(QemuDevice486X8664Cpu{})
+	// gob.Register(QemuDeviceAthlonV1X8664Cpu{})
+	// gob.Register(QemuDeviceAthlonX8664Cpu{})
+	// gob.Register(QemuDeviceBaseX8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellNotsxIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellNotsxX8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellV1X8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellV2X8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellV3X8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellV4X8664Cpu{})
+	// gob.Register(QemuDeviceBroadwellX8664Cpu{})
+	// gob.Register(QemuDeviceCascadelakeServerNotsxX8664Cpu{})
+	// gob.Register(QemuDeviceCascadelakeServerV1X8664Cpu{})
+	// gob.Register(QemuDeviceCascadelakeServerV2X8664Cpu{})
+	// gob.Register(QemuDeviceCascadelakeServerV3X8664Cpu{})
+	// gob.Register(QemuDeviceCascadelakeServerV4X8664Cpu{})
+	// gob.Register(QemuDeviceCascadelakeServerX8664Cpu{})
+	// gob.Register(QemuDeviceConroeV1X8664Cpu{})
+	// gob.Register(QemuDeviceConroeX8664Cpu{})
+	// gob.Register(QemuDeviceCooperlakeV1X8664Cpu{})
+	// gob.Register(QemuDeviceCooperlakeX8664Cpu{})
+	// gob.Register(QemuDeviceCore2duoV1X8664Cpu{})
+	// gob.Register(QemuDeviceCore2duoX8664Cpu{})
+	// gob.Register(QemuDeviceCoreduoV1X8664Cpu{})
+	// gob.Register(QemuDeviceCoreduoX8664Cpu{})
+	// gob.Register(QemuDeviceDenvertonV1X8664Cpu{})
+	// gob.Register(QemuDeviceDenvertonV2X8664Cpu{})
+	// gob.Register(QemuDeviceDenvertonX8664Cpu{})
+	// gob.Register(QemuDeviceDhyanaV1X8664Cpu{})
+	// gob.Register(QemuDeviceDhyanaX8664Cpu{})
+	// gob.Register(QemuDeviceEpycIbpbX8664Cpu{})
+	// gob.Register(QemuDeviceEpycRomeV1X8664Cpu{})
+	// gob.Register(QemuDeviceEpycRomeX8664Cpu{})
+	// gob.Register(QemuDeviceEpycV1X8664Cpu{})
+	// gob.Register(QemuDeviceEpycV2X8664Cpu{})
+	// gob.Register(QemuDeviceEpycV3X8664Cpu{})
+	// gob.Register(QemuDeviceEpycX8664Cpu{})
+	// gob.Register(QemuDeviceHaswellIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceHaswellNotsxIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceHaswellNotsxX8664Cpu{})
+	// gob.Register(QemuDeviceHaswellV1X8664Cpu{})
+	// gob.Register(QemuDeviceHaswellV2X8664Cpu{})
+	// gob.Register(QemuDeviceHaswellV3X8664Cpu{})
+	// gob.Register(QemuDeviceHaswellV4X8664Cpu{})
+	// gob.Register(QemuDeviceHaswellX8664Cpu{})
+	// gob.Register(QemuDeviceHostX8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeClientNotsxX8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeClientV1X8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeClientV2X8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeClientX8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeServerNotsxX8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeServerV1X8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeServerV2X8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeServerV3X8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeServerV4X8664Cpu{})
+	// gob.Register(QemuDeviceIcelakeServerX8664Cpu{})
+	// gob.Register(QemuDeviceIvybridgeIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceIvybridgeV1X8664Cpu{})
+	// gob.Register(QemuDeviceIvybridgeV2X8664Cpu{})
+	// gob.Register(QemuDeviceIvybridgeX8664Cpu{})
+	// gob.Register(QemuDeviceKnightsmillV1X8664Cpu{})
+	// gob.Register(QemuDeviceKnightsmillX8664Cpu{})
+	// gob.Register(QemuDeviceKvm32V1X8664Cpu{})
+	// gob.Register(QemuDeviceKvm32X8664Cpu{})
+	// gob.Register(QemuDeviceKvm64V1X8664Cpu{})
+	// gob.Register(QemuDeviceKvm64X8664Cpu{})
+	// gob.Register(QemuDeviceMaxX8664Cpu{})
+	// gob.Register(QemuDeviceN270V1X8664Cpu{})
+	// gob.Register(QemuDeviceN270X8664Cpu{})
+	// gob.Register(QemuDeviceNehalemIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceNehalemV1X8664Cpu{})
+	// gob.Register(QemuDeviceNehalemV2X8664Cpu{})
+	// gob.Register(QemuDeviceNehalemX8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG1V1X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG1X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG2V1X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG2X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG3V1X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG3X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG4V1X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG4X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG5V1X8664Cpu{})
+	// gob.Register(QemuDeviceOpteronG5X8664Cpu{})
+	// gob.Register(QemuDevicePenrynV1X8664Cpu{})
+	// gob.Register(QemuDevicePenrynX8664Cpu{})
+	// gob.Register(QemuDevicePentiumV1X8664Cpu{})
+	// gob.Register(QemuDevicePentiumX8664Cpu{})
+	// gob.Register(QemuDevicePentium2V1X8664Cpu{})
+	// gob.Register(QemuDevicePentium2X8664Cpu{})
+	// gob.Register(QemuDevicePentium3V1X8664Cpu{})
+	// gob.Register(QemuDevicePentium3X8664Cpu{})
+	// gob.Register(QemuDevicePhenomV1X8664Cpu{})
+	// gob.Register(QemuDevicePhenomX8664Cpu{})
+	// gob.Register(QemuDeviceQemu32V1X8664Cpu{})
+	// gob.Register(QemuDeviceQemu32X8664Cpu{})
+	// gob.Register(QemuDeviceQemu64V1X8664Cpu{})
+	// gob.Register(QemuDeviceQemu64X8664Cpu{})
+	// gob.Register(QemuDeviceSandybridgeIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceSandybridgeV1X8664Cpu{})
+	// gob.Register(QemuDeviceSandybridgeV2X8664Cpu{})
+	// gob.Register(QemuDeviceSandybridgeX8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeClientIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeClientNotsxIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeClientV1X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeClientV2X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeClientV3X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeClientX8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerNotsxIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerV1X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerV2X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerV3X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerV4X8664Cpu{})
+	// gob.Register(QemuDeviceSkylakeServerX8664Cpu{})
+	// gob.Register(QemuDeviceSnowridgeV1X8664Cpu{})
+	// gob.Register(QemuDeviceSnowridgeV2X8664Cpu{})
+	// gob.Register(QemuDeviceSnowridgeX8664Cpu{})
+	// gob.Register(QemuDeviceWestmereIbrsX8664Cpu{})
+	// gob.Register(QemuDeviceWestmereV1X8664Cpu{})
+	// gob.Register(QemuDeviceWestmereV2X8664Cpu{})
+	// gob.Register(QemuDeviceWestmereX8664Cpu{})
+
+	// Controller/Bridge/Hub devices
+	// gob.Register(QemuDeviceI82801b11Bridge{})
+	// gob.Register(QemuDeviceIgdPassthroughIsaBridge{})
+	// gob.Register(QemuDeviceIoh3420{})
+	// gob.Register(QemuDevicePciBridge{})
+	// gob.Register(QemuDevicePciBridgeSeat{})
+	// gob.Register(QemuDevicePciePciBridge{})
+	// gob.Register(QemuDevicePcieRootPort{})
+	// gob.Register(QemuDevicePxb{})
+	// gob.Register(QemuDevicePxbPcie{})
+	// gob.Register(QemuDeviceUsbHost{})
+	// gob.Register(QemuDeviceUsbHub{})
+	// gob.Register(QemuDeviceVfioPciIgdLpcBridge{})
+	// gob.Register(QemuDeviceVmbusBridge{})
+	// gob.Register(QemuDeviceX3130Upstream{})
+	// gob.Register(QemuDeviceXio3130Downstream{})
+
+	// Display devices
+	// gob.Register(QemuDeviceAtiVga{})
+	// gob.Register(QemuDeviceBochsDisplay{})
+	// gob.Register(QemuDeviceCirrusVga{})
+	// gob.Register(QemuDeviceIsaCirrusVga{})
+	// gob.Register(QemuDeviceIsaVga{})
+	// gob.Register(QemuDeviceQxl{})
+	// gob.Register(QemuDeviceQxlVga{})
+	// gob.Register(QemuDeviceRamfb{})
+	// gob.Register(QemuDeviceSecondaryVga{})
+	gob.Register(QemuDeviceSga{})
+	// gob.Register(QemuDeviceVga{})
+	// gob.Register(QemuDeviceVhostUserGpu{})
+	// gob.Register(QemuDeviceVhostUserGpuPci{})
+	// gob.Register(QemuDeviceVhostUserVga{})
+	// gob.Register(QemuDeviceVirtioGpuDevice{})
+	// gob.Register(QemuDeviceVirtioGpuPci{})
+	// gob.Register(QemuDeviceVirtioVga{})
+	// gob.Register(QemuDeviceVmwareSvga{})
+
+	// Input devices
+	// gob.Register(QemuDeviceCcidCardEmulated{})
+	// gob.Register(QemuDeviceCcidCardPassthru{})
+	// gob.Register(QemuDeviceI8042{})
+	// gob.Register(QemuDeviceIpoctal232{})
+	// gob.Register(QemuDeviceIsaParallel{})
+	// gob.Register(QemuDeviceIsaSerial{})
+	// gob.Register(QemuDevicePciSerial{})
+	// gob.Register(QemuDevicePciSerial2x{})
+	// gob.Register(QemuDevicePciSerial4x{})
+	// gob.Register(QemuDeviceTpci200{})
+	// gob.Register(QemuDeviceUsbBraille{})
+	// gob.Register(QemuDeviceUsbCcid{})
+	// gob.Register(QemuDeviceUsbKbd{})
+	// gob.Register(QemuDeviceUsbMouse{})
+	// gob.Register(QemuDeviceUsbSerial{})
+	// gob.Register(QemuDeviceUsbTablet{})
+	// gob.Register(QemuDeviceUsbWacomTablet{})
+	// gob.Register(QemuDeviceVhostUserInput{})
+	// gob.Register(QemuDeviceVhostUserInputPci{})
+	// gob.Register(QemuDeviceVirtconsole{})
+	// gob.Register(QemuDeviceVirtioInputHostDevice{})
+	// gob.Register(QemuDeviceVirtioInputHostPci{})
+	// gob.Register(QemuDeviceVirtioKeyboardDevice{})
+	// gob.Register(QemuDeviceVirtioKeyboardPci{})
+	// gob.Register(QemuDeviceVirtioMouseDevice{})
+	// gob.Register(QemuDeviceVirtioMousePci{})
+	// gob.Register(QemuDeviceVirtioSerialDevice{})
+	// gob.Register(QemuDeviceVirtioSerialPci{})
+	// gob.Register(QemuDeviceVirtioSerialPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioSerialPciTransitional{})
+	// gob.Register(QemuDeviceVirtioTabletDevice{})
+	// gob.Register(QemuDeviceVirtioTabletPci{})
+	// gob.Register(QemuDeviceVirtserialport{})
+
+	// Misc devices
+	// gob.Register(QemuDeviceAmdIommu{})
+	// gob.Register(QemuDeviceCtucanPci{})
+	// gob.Register(QemuDeviceEdu{})
+	// gob.Register(QemuDeviceHypervTestdev{})
+	// gob.Register(QemuDeviceI2cDdc{})
+	// gob.Register(QemuDeviceI6300esb{})
+	// gob.Register(QemuDeviceIb700{})
+	// gob.Register(QemuDeviceIntelIommu{})
+	// gob.Register(QemuDeviceIsaApplesmc{})
+	// gob.Register(QemuDeviceIsaDebugExit{})
+	// gob.Register(QemuDeviceIsaDebugcon{})
+	// gob.Register(QemuDeviceIvshmemDoorbell{})
+	// gob.Register(QemuDeviceIvshmemPlain{})
+	// gob.Register(QemuDeviceKvaserPci{})
+	// gob.Register(QemuDeviceLoader{})
+	// gob.Register(QemuDeviceMioe3680Pci{})
+	// gob.Register(QemuDevicePcTestdev{})
+	// gob.Register(QemuDevicePciTestdev{})
+	// gob.Register(QemuDevicePcm3680Pci{})
+	// gob.Register(QemuDevicePvpanic{})
+	// gob.Register(QemuDeviceSmbusIpmi{})
+	// gob.Register(QemuDeviceTpmCrb{})
+	// gob.Register(QemuDeviceUsbRedir{})
+	// gob.Register(QemuDeviceVfioPci{})
+	// gob.Register(QemuDeviceVfioPciNohotplug{})
+	// gob.Register(QemuDeviceVhostUserVsockDevice{})
+	// gob.Register(QemuDeviceVhostUserVsockPci{})
+	// gob.Register(QemuDeviceVhostUserVsockPciNonTransitional{})
+	// gob.Register(QemuDeviceVhostVsockDevice{})
+	// gob.Register(QemuDeviceVhostVsockPci{})
+	// gob.Register(QemuDeviceVhostVsockPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioBalloonDevice{})
+	// gob.Register(QemuDeviceVirtioBalloonPci{})
+	// gob.Register(QemuDeviceVirtioBalloonPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioBalloonPciTransitional{})
+	// gob.Register(QemuDeviceVirtioCryptoDevice{})
+	// gob.Register(QemuDeviceVirtioCryptoPci{})
+	// gob.Register(QemuDeviceVirtioIommuDevice{})
+	// gob.Register(QemuDeviceVirtioIommuPci{})
+	// gob.Register(QemuDeviceVirtioIommuPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioMem{})
+	// gob.Register(QemuDeviceVirtioMemPci{})
+	// gob.Register(QemuDeviceVirtioPmemPci{})
+	// gob.Register(QemuDeviceVirtioRngDevice{})
+	// gob.Register(QemuDeviceVirtioRngPci{})
+	// gob.Register(QemuDeviceVirtioRngPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioRngPciTransitional{})
+	// gob.Register(QemuDeviceVmcoreinfo{})
+	// gob.Register(QemuDeviceVmgenid{})
+	// gob.Register(QemuDeviceXenBackend{})
+	// gob.Register(QemuDeviceXenPciPassthrough{})
+	// gob.Register(QemuDeviceXenPlatform{})
+
+	// Network devices
+	// gob.Register(QemuDeviceE1000{})
+	// gob.Register(QemuDeviceE100082544gc{})
+	// gob.Register(QemuDeviceE100082545em{})
+	// gob.Register(QemuDeviceE1000e{})
+	// gob.Register(QemuDeviceI82550{})
+	// gob.Register(QemuDeviceI82551{})
+	// gob.Register(QemuDeviceI82557a{})
+	// gob.Register(QemuDeviceI82557b{})
+	// gob.Register(QemuDeviceI82557c{})
+	// gob.Register(QemuDeviceI82558a{})
+	// gob.Register(QemuDeviceI82558b{})
+	// gob.Register(QemuDeviceI82559a{})
+	// gob.Register(QemuDeviceI82559b{})
+	// gob.Register(QemuDeviceI82559c{})
+	// gob.Register(QemuDeviceI82559er{})
+	// gob.Register(QemuDeviceI82562{})
+	// gob.Register(QemuDeviceI82801{})
+	// gob.Register(QemuDeviceNe2kIsa{})
+	// gob.Register(QemuDeviceNe2kPci{})
+	// gob.Register(QemuDevicePcnet{})
+	// gob.Register(QemuDevicePvrdma{})
+	// gob.Register(QemuDeviceRocker{})
+	// gob.Register(QemuDeviceRtl8139{})
+	// gob.Register(QemuDeviceTulip{})
+	// gob.Register(QemuDeviceUsbNet{})
+	// gob.Register(QemuDeviceVirtioNetDevice{})
+	// gob.Register(QemuDeviceVirtioNetPci{})
+	// gob.Register(QemuDeviceVirtioNetPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioNetPciTransitional{})
+	// gob.Register(QemuDeviceVmxnet3{})
+
+	// Sound devices
+	// gob.Register(QemuDeviceAc97{})
+	// gob.Register(QemuDeviceAdlib{})
+	// gob.Register(QemuDeviceCs4231a{})
+	// gob.Register(QemuDeviceEs1370{})
+	// gob.Register(QemuDeviceGus{})
+	// gob.Register(QemuDeviceHdaDuplex{})
+	// gob.Register(QemuDeviceHdaMicro{})
+	// gob.Register(QemuDeviceHdaOutput{})
+	// gob.Register(QemuDeviceIch9IntelHda{})
+	// gob.Register(QemuDeviceIntelHda{})
+	// gob.Register(QemuDeviceSb16{})
+	// gob.Register(QemuDeviceUsbAudio{})
+
+	// Storage devices
+	// gob.Register(QemuDeviceAm53c974{})
+	// gob.Register(QemuDeviceDc390{})
+	// gob.Register(QemuDeviceFloppy{})
+	// gob.Register(QemuDeviceIch9Ahci{})
+	// gob.Register(QemuDeviceIdeCd{})
+	// gob.Register(QemuDeviceIdeDrive{})
+	// gob.Register(QemuDeviceIdeHd{})
+	// gob.Register(QemuDeviceIsaFdc{})
+	// gob.Register(QemuDeviceIsaIde{})
+	// gob.Register(QemuDeviceLsi53c810{})
+	// gob.Register(QemuDeviceLsi53c895a{})
+	// gob.Register(QemuDeviceMegasas{})
+	// gob.Register(QemuDeviceMegasasGen2{})
+	// gob.Register(QemuDeviceMptsas1068{})
+	// gob.Register(QemuDeviceNvme{})
+	// gob.Register(QemuDeviceNvmeNs{})
+	// gob.Register(QemuDevicePiix3Ide{})
+	// gob.Register(QemuDevicePiix3IdeXen{})
+	// gob.Register(QemuDevicePiix4Ide{})
+	// gob.Register(QemuDevicePvscsi{})
+	// gob.Register(QemuDeviceScsiBlock{})
+	// gob.Register(QemuDeviceScsiCd{})
+	// gob.Register(QemuDeviceScsiDisk{})
+	// gob.Register(QemuDeviceScsiGeneric{})
+	// gob.Register(QemuDeviceScsiHd{})
+	// gob.Register(QemuDeviceSdCard{})
+	// gob.Register(QemuDeviceSdhciPci{})
+	// gob.Register(QemuDeviceUsbBot{})
+	// gob.Register(QemuDeviceUsbMtp{})
+	// gob.Register(QemuDeviceUsbStorage{})
+	// gob.Register(QemuDeviceUsbUas{})
+	// gob.Register(QemuDeviceVhostScsi{})
+	// gob.Register(QemuDeviceVhostScsiPci{})
+	// gob.Register(QemuDeviceVhostScsiPciNonTransitional{})
+	// gob.Register(QemuDeviceVhostScsiPciTransitional{})
+	// gob.Register(QemuDeviceVhostUserBlk{})
+	// gob.Register(QemuDeviceVhostUserBlkPci{})
+	// gob.Register(QemuDeviceVhostUserBlkPciNonTransitional{})
+	// gob.Register(QemuDeviceVhostUserBlkPciTransitional{})
+	// gob.Register(QemuDeviceVhostUserFsDevice{})
+	// gob.Register(QemuDeviceVhostUserFsPci{})
+	// gob.Register(QemuDeviceVhostUserScsi{})
+	// gob.Register(QemuDeviceVhostUserScsiPci{})
+	// gob.Register(QemuDeviceVhostUserScsiPciNonTransitional{})
+	// gob.Register(QemuDeviceVhostUserScsiPciTransitional{})
+	// gob.Register(QemuDeviceVirtio9pDevice{})
+	// gob.Register(QemuDeviceVirtio9pPci{})
+	// gob.Register(QemuDeviceVirtio9pPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtio9pPciTransitional{})
+	// gob.Register(QemuDeviceVirtioBlkDevice{})
+	// gob.Register(QemuDeviceVirtioBlkPci{})
+	// gob.Register(QemuDeviceVirtioBlkPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioBlkPciTransitional{})
+	// gob.Register(QemuDeviceVirtioScsiDevice{})
+	// gob.Register(QemuDeviceVirtioScsiPci{})
+	// gob.Register(QemuDeviceVirtioScsiPciNonTransitional{})
+	// gob.Register(QemuDeviceVirtioScsiPciTransitional{})
+
+	// USB devices
+	// gob.Register(QemuDeviceIch9UsbEhci1{})
+	// gob.Register(QemuDeviceIch9UsbEhci2{})
+	// gob.Register(QemuDeviceIch9UsbUhci1{})
+	// gob.Register(QemuDeviceIch9UsbUhci2{})
+	// gob.Register(QemuDeviceIch9UsbUhci3{})
+	// gob.Register(QemuDeviceIch9UsbUhci4{})
+	// gob.Register(QemuDeviceIch9UsbUhci5{})
+	// gob.Register(QemuDeviceIch9UsbUhci6{})
+	// gob.Register(QemuDeviceNecUsbXhci{})
+	// gob.Register(QemuDevicePciOhci{})
+	// gob.Register(QemuDevicePiix3UsbUhci{})
+	// gob.Register(QemuDevicePiix4UsbUhci{})
+	// gob.Register(QemuDeviceQemuXhci{})
+	// gob.Register(QemuDeviceUsbEhci{})
+	// gob.Register(QemuDeviceVt82c686bUsbUhci{})
+
+	// Uncategorized devices
+	// gob.Register(QemuDeviceAmdviPci{})
+	// gob.Register(QemuDeviceIpmiBmcExtern{})
+	// gob.Register(QemuDeviceIpmiBmcSim{})
+	// gob.Register(QemuDeviceIsaIpmiBt{})
+	// gob.Register(QemuDeviceIsaIpmiKcs{})
+	// gob.Register(QemuDeviceMc146818rtc{})
+	// gob.Register(QemuDeviceNvdimm{})
+	// gob.Register(QemuDevicePcDimm{})
+	// gob.Register(QemuDevicePciIpmiBt{})
+	// gob.Register(QemuDevicePciIpmiKcs{})
+	// gob.Register(QemuDeviceTpmTis{})
+	// gob.Register(QemuDeviceU2fPassthru{})
+	// gob.Register(QemuDeviceVirtioPmem{})
+	// gob.Register(QemuDeviceVmmouse{})
+	// gob.Register(QemuDeviceXenCdrom{})
+	// gob.Register(QemuDeviceXenDisk{})
+	// gob.Register(QemuDeviceXenPvdevice{})
+
+	// CPUs
+	gob.Register(QemuCPU{})
+	gob.Register(QemuCPUX86(""))
+	gob.Register(QemuCPUArm(""))
+
+	// Displays
+	// gob.Register(QemuDisplaySpiceApp{})
+	// gob.Register(QemuDisplayGtk{})
+	// gob.Register(QemuDisplayVNC{})
+	// gob.Register(QemuDisplayCurses{})
+	// gob.Register(QemuDisplayEglHeadless{})
+	gob.Register(QemuDisplayNone{})
+}
